fix: avoid flickering when loading initial entry
